feat(RulesDetail): display of decision rule execution errors
