+++ packages/app-builder/src/components/Scenario/AstBuilder/ErrorMessage.tsx
@@ -1,4 +1,4 @@
-import { type EvaluationError } from '@app-builder/models';
+import { type EvaluationError } from '@app-builder/models/node-evaluation';
 import {
   adaptEvaluationErrorViewModels,
   useGetNodeEvaluationErrorMessage,
