+++ packages/app-builder/src/services/validation/scenario-validation.ts
@@ -1,8 +1,8 @@
+import { type ScenarioValidation } from '@app-builder/models';
 import {
   type EvaluationError,
   type NodeEvaluation,
-  type ScenarioValidation,
-} from '@app-builder/models';
+} from '@app-builder/models/node-evaluation';
 import invariant from 'tiny-invariant';
 
 // return just an array of error from a recursive evaluation
